Move `TestCudaComm` to multigpu
